Drop Hadrian.Package.Type
